Introduce defcommand as a key->function binding
Previously the implementation required including the actual function
using the function-quote syntax and then putting this mapping into a
lookup table. This meant that the lookup table was relatively static.

Now the names of functions are used instead and I introduced defcommand.

`defcommand` will create a command table itself and take care of lookups
as well, similar to the previous implementation. The difference here is
that the table can be changed without restarting Lispkit.